requirements.txt
appdirs==1.4.3
astroid==2.3.3
attrs==19.3.0
black==19.10b0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
codecov==2.0.22
coloredlogs==14.0
coverage==5.0.4
humanfriendly==8.1
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.3
pathspec==0.7.0
pluggy==0.13.1
py==1.8.1
pycodestyle==2.5.0
pylint==2.4.4
pyparsing==2.4.6
pytest==5.4.1
pytest-cov==2.8.1
PyYAML==5.3.1
regex==2020.2.20
requests==2.23.0
six==1.14.0
termcolor==1.1.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.8
wcwidth==0.1.8
wrapt==1.11.2
